requirements.txt
asgiref==3.7.2
attrs==23.2.0
build==0.10.0
Django==5.0.1
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==22.5.0
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==23.1
pluggy==1.3.0
pyproject_hooks==1.0.0
pytest==7.4.4
pytest-django==4.7.0
python-dateutil==2.8.2
referencing==0.32.1
rpds-py==0.17.1
setuptools-scm==7.1.0
-e git+ssh://git@github.com/Aiky30/shed-pi.git@054fbb18219482fa4bb7bc4c2c883941fb647def#egg=shed_pi
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
